backend/requirements.txt
fastapi
uvicorn
sqlalchemy
mysql-connector-python
passlib[bcrypt]
python-jose
openai
boto3
python-dotenv